adding preset mode to system

+++ main/global_state.h
@@ -151,7 +151,7 @@ typedef struct
     uint32_t chip_submit[8];
     char chip_submit_srt[128];
 
-    uint8_t runningMode;
+    uint16_t runningMode;
 
     TaskHandle_t job_task_h;
     TaskHandle_t asic_task_h;
